generate command extraction
