+++ vscode/microsoft-kiota/src/commands/GeneratedOutputState.ts
@@ -1,0 +1,4 @@
+export interface GeneratedOutputState {
+  outputPath: string;
+  clientClassName: string;
+}
